requirements.txt
alabaster==0.7.12
asyncio==3.4.3
atomicwrites==1.3.0
attrdict==2.0.1
attrs==19.1.0
Babel==2.7.0
base58==1.0.3
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configparser==3.5.0
contextlib2==0.5.5
cytoolz==0.10.0
docutils==0.15.2
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.4
hexbytes==0.2.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
ipfshttpclient==0.4.12
Jinja2==2.10.1
jsonschema==2.6.0
lru-dict==1.1.6
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==7.2.0
multiaddr==0.0.8
netaddr==0.7.19
numpy==1.17.0
packaging==19.1
parsimonious==0.8.1
pkg-resources==0.0.0
pluggy==0.12.0
protobuf==3.9.1
py==1.8.0
py-geth==2.1.0
py-solc==3.2.0
pycryptodome==3.8.2
Pygments==2.4.2
pyparsing==2.4.2
pytest==4.6.5
pytz==2019.2
requests==2.22.0
rlp==1.1.0
semantic-version==2.6.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinx-better-theme==0.1.5
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
toolz==0.10.0
toposort==1.5
urllib3==1.25.3
varint==1.0.2
wcwidth==0.1.7
-e git+https://github.com/ethereum/web3.py.git@01380b5f4c32115c59390b91efe602c0b40b44a1#egg=web3
websockets==7.0
zipp==0.5.2
